backend/requirements.txt
flask
flask-cors
openai-whisper
transformers
torch
faiss-cpu
sentence-transformers
python-dotenv
googletrans==4.0.0rc1
pydub